+++ Ares/src/Ares/Renderer/Renderer2D.cpp
@@ -5,7 +5,8 @@
 
 #include "Ares/Renderer/Renderer.h"
 
-#include <glm/gtc/matrix_transform.hpp>
+//#include <glm/gtc/matrix_transform.hpp>
+#include "Ares/Math/Math.h"
 
 namespace Ares 
 {
@@ -253,7 +254,7 @@ namespace Ares
 			s_Data.Stats.DrawCalls++;	
 		}
 
-		dataSize = (uint8_t*)s_Data.LineVertexBufferPtr - (uint8_t*)s_Data.LineVertexBufferBase;
+		dataSize = (uint32_t)((uint8_t*)s_Data.LineVertexBufferPtr - (uint8_t*)s_Data.LineVertexBufferBase);
 		if (dataSize)
 		{
 			s_Data.LineVertexBuffer->SetData(s_Data.LineVertexBufferBase, dataSize);
